Update udev rules for new RFD900 radio (#50)
Update the udev rules so that the onboard computer can disambiguate between the rfd900 and dji sdk interface. Add separate set of udev rules for the ground control station since the attributes for the GCS radio are slightly different. Note, these rules may need to be changed if the radio usb to ftdi cable is changed. These rules have been tested onboard and on the GCS.