Don't send DJI SDK commands if RC not in F-mode (#54)
Modify the DJI SDK ROS wrapper to check if the RC is in F-mode. If not, don't send any flight control commands to the DJI flight controller. Add a publisher for the control authority status (1 if the SDK has been granted control authority, 0 otherwise). If the control authority status is 1, but the RC is not in F-mode, release control authority. Add a bug fix to the onboard telemetry.