Remove replanning from firefly, let IPP do it. Note this means that IPP will not wait for us to execute the first plan, but rather keep replanning right away. We should come back to fix this later
The reason why we did it this way was because sending a Replan message from Firefly was causing the map to clear and the drone to go back to its first waypoint for some reason. I think this is a bug in the ipp_planners_node.h. We will fix it later